+++ multiqc/modules/ngsbits/__init__.py
@@ -1,0 +1,3 @@
+from .ngsbits import MultiqcModule
+
+__all__ = ["MultiqcModule"]